Automation! For doing 3 tests at once
To automate the testing process, we are now sending 3 images per test to do 3 tests at once. 256modo is the vector file with 3 images sparated with a 256 digit between image one and two and between two and three.  In addition to this LongVectorAnalysis.m is an updated version of the old vector analysis code, it chops the new vectors into 3 separate images to process the new data transferring method and does all the relevant analysis computations. Also some other minor updates to the arduino code to pad out the last few hardware caused errors when transferring files.